Add files via upload
This is what I have for a 'cleanup' so far. The first chunk of code replaces all the double-whitespaces with single-whitespaces. The second chunk of code deletes everything before the detailed description (we don't need nothing before it) by placing everything after and including the Detailed Description in a new txt file.