Add files via upload
This IS NOT the fully fledged comparison algo. So far, this algo can tell you whether or not the lines are identical. The algo compares the parent file against the CIP. If the algo detects a new line that isn't in the parent file, it will stop iterating over the parent lines of text and continue to iterate over the CIP lines of text UNTIL it finds an identical line of text to the current line of the parent.